freeciv: fix dependencies
Freeciv has been depending on gtk+; however, it would not actually
build the gtk+ client. Instead it would try to pick up the SDL
libraries, and fail to build altogether if SDL was not installed.

This removes the gtk+ dep and adds SDL, along with a dependency on :x11
(for freetype).

Patches which fix building with gtk+ are welcome!

Fixes #13.